Conflicts:
	coreo/templates/ucore/map.html
	coreo/ucore/views.py

+++ coreo/media/javascript/core.ui/geodatatree.js
@@ -14,6 +14,7 @@
  *   - core.geo.GeoDataStore
  *   - core.util.Assert
  *   - core.geo.KmlFeatureType
+ *   - core.util.GeoDataVisitor
  */
 
 if (!window.core)
@@ -26,8 +27,11 @@ if (!window.core.ui)
 	var Assert = core.util.Assert;
 	var KmlFeatureType = core.geo.KmlFeatureType;
 	if (!KmlFeatureType)
-		throw ("Dependency not found: core.geo.KmlFeatureType");
-	
+		throw "Dependency not found: core.geo.KmlFeatureType";
+	var GeoDataVisitor = core.util.GeoDataVisitor;
+	if (!GeoDataVisitor)
+		throw "Dependency not found: core.util.GeoDataVisitor";
+
 	/**
 	 * Constructor: GeoDataTree
 	 * 
@@ -37,13 +41,14 @@ if (!window.core.ui)
 	 *   geodata - <GeoData>. Required. Root tree node.
 	 *   el - HTML DOM Element. Element to contain the tree.
 	 */
-	var GeoDataTree = function(name, geodata, el, networkLinkQueue) {
+	var GeoDataTree = function(name, geodata, el, networkLinkQueue, showEditButton) {
 		Assert.notNull(geodata, "geodata cannot be null");
 		Assert.type(geodata, "object", "geodata must be of type core.geo.GeoData");
 		Assert.notNull(el, "el cannot be null");
 		this.el = el;
 		this.geodata = geodata;
 		this.networkLinkQueue = networkLinkQueue;
+		this.showEditButton = showEditButton;
 		this._init(name);
 	};
 	/**
@@ -53,6 +58,7 @@ if (!window.core.ui)
 	 */
 	GeoDataTree.GEODATA_ATTR = "core-geodata-id";
 	GeoDataTree.KML_TYPE_ATTR = "core-kml-type";
+	GeoDataTree.CORE_NODE_TYPE_ATTR = "core-node-type";
 	GeoDataTree.prototype = {
 		
 		/**
@@ -123,13 +129,25 @@ if (!window.core.ui)
 		 * Function. Invoked after a geodata is removed.
 		 */
 		onRemove: function(geodata) {},
-		
+
 		/**
-		 * Property: onRename
+		 * Property: onEdit
 		 * 
-		 * Function. Invoked after a geodata is renamed.
+		 * Function. Invoked when an edit button is clicked.
 		 */
-		onRename: function(geodata, newName) {},
+		onEdit: function(geodata) {},
+
+		deselectAll: function() {
+			$(this.el).jstree("deselect_all");
+		},
+
+		isCoreLink: function(geodata) {
+			return geodata && typeof geodata.getCoreLink === "function";
+		},
+		
+		isCoreLinkLibrary: function(geodata) {
+			return geodata && typeof geodata.getLinkLibrary === "function";
+		},
 		
 		/**
 		 * Function: _createTreeNode
@@ -155,8 +173,12 @@ if (!window.core.ui)
 				name = geodata.getKmlFeatureType();
 			}
 			
+			var nodeType = this.isCoreLink(geodata) ? "core-link"
+					: this.isCoreLinkLibrary(geodata) ? "core-link-library"
+							: geodata.getKmlFeatureType();
+				
 			var title = $("<span>")
-						.append($("<ins>").addClass(geodata.getKmlFeatureType()).html("&#160;"))
+						.append($("<ins>").addClass(nodeType).html("&#160;"))
 						.append($("<span>").html(name));
 			if (this.appendHoverActions && typeof this.appendHoverActions === "function") {
 				var hoverActions = $("<div>").addClass("geodatatree-hoveractions");
@@ -167,6 +189,17 @@ if (!window.core.ui)
 			treeNode.attr = {};
 			treeNode.attr[GeoDataTree.GEODATA_ATTR] = geodata.id;
 			treeNode.attr[GeoDataTree.KML_TYPE_ATTR] = geodata.getKmlFeatureType();
+			new GeoDataVisitor({
+				link: function() {
+					treeNode["class"] = "jstree-draggable";
+					treeNode.attr[GeoDataTree.CORE_NODE_TYPE_ATTR] = "Link";
+				},
+				linkLibrary: function() {
+					treeNode["class"] = "jstree-drop";
+					treeNode.attr[GeoDataTree.CORE_NODE_TYPE_ATTR] = "LinkLibrary";					
+				}
+			}).visit(geodata);
+			
 			if (geodata.hasChildren()) {
 				treeNode.state = "closed";
 			}
@@ -175,6 +208,12 @@ if (!window.core.ui)
 			}
 			// treeNode.icon = geodata.getKmlFeatureType();
 			return treeNode;
+		},
+
+		rename: function(geoDataId, newName) {
+			var treeNode = this._findNodeByGeoDataId(geoDataId);
+			treeNode.find("> a:eq(1) > span > span")
+				.html(newName);
 		},
 
 		/**
@@ -220,13 +259,23 @@ if (!window.core.ui)
 					}
 				}
 			})
-			.bind("select_node.jstree", function(e, data) {
+			.bind("select_node.jstree", $.proxy(function(e, data) {
+				$(this.el).find("li.ui-state-active")
+					.removeClass("ui-state-active");
 				var selected = data.rslt.obj;
+				selected.addClass("ui-state-active");
 				var geodata = getGeoDataFromTreeNode(selected);
 				if (_this.onSelect) {
 					_this.onSelect(geodata);
 				}
+			}, this))
+			.bind("deselect_node.jstree", function(e, data) {
+				var deselected = data.rslt.obj;
+				deselected.removeClass("ui-state-active");
 			})
+			.bind("deselect_all.jstree", $.proxy(function(e, data) {
+				$(this.el).find("li.ui-state-active").removeClass("ui-state-active");
+			}, this))
 			.bind("hover_node.jstree", function(e, data) {
 				var hovered = data.rslt.obj;
 				var geodata = getGeoDataFromTreeNode(hovered);
@@ -242,10 +291,12 @@ if (!window.core.ui)
 				}
 			})
 			.bind("remove.jstree", function(e, data) {
-				var removed = data.rslt.obj;
-				var geodata = getGeoDataFromTreeNode(removed);
-				if (_this.onRemove) {
-					_this.onRemove(geodata);
+				var removed = data && data.rslt && data.rslt.obj ? data.rslt.obj : undefined;
+				if (removed) {
+					var geodata = getGeoDataFromTreeNode(removed);
+					if (_this.onRemove) {
+						_this.onRemove(geodata);
+					}
 				}
 			})
 			.bind("rename.jstree", function(e, data) {
@@ -302,53 +353,18 @@ if (!window.core.ui)
 					getHoverButtons: function(node) {
 						var kmlType = node.attr(GeoDataTree.KML_TYPE_ATTR);
 						var buttons = [];
-						if (kmlType === KmlFeatureType.FOLDER 
-								|| kmlType === KmlFeatureType.DOCUMENT) {
+						if (_this.showEditButton) {
 							buttons.push({ 
-					        	icon: "ui-icon-plusthick", 
-					        	tooltip: "Add folder", 
-					        	action: function() {
-					        		// TODO: Create geodata, then create tree node from it
-					        		var newNode = {
-					        			attr: {},
-					        			state: "closed",
-					        			data: [
-					        			       {
-					        			    	   title: $("<span>")
-							    						.append($("<ins>").addClass(KmlFeatureType.FOLDER).html("&#160;"))
-							    						.append($("<span>").text("New Folder")),
-							    					icon: KmlFeatureType.FOLDER
-					        			       }
-					        			]
-					        		};
-					        		newNode.attr[GeoDataTree.KML_TYPE_ATTR] = KmlFeatureType.FOLDER;
-					        		
-//					    			var title = $("<span>")
-//					    						.append($("<ins>").addClass(KmlFeatureType.FOLDER).html("&#160;"))
-//					    						.append($("<span>").text("New Folder"));
-//					    			if (this.appendHoverActions && typeof this.appendHoverActions === "function") {
-//					    				var hoverActions = $("<div>").addClass("geodatatree-hoveractions");
-//					    				this.appendHoverActions.call(this.appendHoverActions, hoverActions, geodata);
-//					    				title.append(hoverActions);
-//					    			}
-//					    			newNode.data
-//					    			newNode.title = title;
-//					    			newNode.attr = {};
-					    			// TODO
-					    			// treeNode.attr[GeoDataTree.GEODATA_ATTR] = geodata.id;
-					    			//newNode.attr[GeoDataTree.KML_TYPE_ATTR] = KmlFeatureType.FOLDER;
-					    			//newNode.state = "closed";
-					        		$(_this.el).jstree("create", node, "first", newNode); 
+					        	icon: "ui-icon-pencil", 
+					        	tooltip: "Edit", 
+					        	action: function(node) {
+					        		if (_this.onEdit) {
+					        			var geodata = getGeoDataFromTreeNode(node);
+					        			_this.onEdit.call(_this.onEdit, geodata);
+					        		}
 					        	} 
 					        });
 						}
-						buttons.push({ 
-				        	icon: "ui-icon-pencil", 
-				        	tooltip: "Rename", 
-				        	action: function(node) {
-				        		$(_this.el).jstree("rename", node);
-				        	} 
-				        });
 						buttons.push({ 
 				        	icon: "ui-icon-trash", 
 				        	tooltip: "Remove", 
@@ -360,26 +376,65 @@ if (!window.core.ui)
 					}
 				},
 				dnd: {
-					drop_target: ".jstree li[" + GeoDataTree.KML_TYPE_ATTR + "='" + KmlFeatureType.FOLDER + "'],"
-								+ ".jstree li[" + GeoDataTree.KML_TYPE_ATTR + "='" + KmlFeatureType.DOCUMENT + "']",
-					drop_check: function(data) {
-						console.log("drop_check");
-						var dragged = data.o;
-						var dropTarget = data.r;
-						var geodata = getGeoDataFromTreeNode(dropTarget);
-						return (geodata.getKmlFeatureType() === KmlFeatureType.DOCUMENT
-								|| geodata.getKmlFeatureType() === KmlFeatureType.FOLDER);
-					},
-					drag_check: function(data) {
-						console.log("drag_check");
-						return {
-							before: false,
-							after: false,
-							inside: false
-						};
+					drop_target: false,
+					drag_target: false
+				},
+				crrm: {
+					move: {
+						always_copy: "multitree",
+						check_move: function(m) {
+							var hovered, dragged, valid;
+							hovered = m.r.closest("li");
+							dragged = m.o.closest("li");
+							valid = (dragged.is("[" + GeoDataTree.CORE_NODE_TYPE_ATTR + "='Link']")
+									&& ((hovered.is("[" + GeoDataTree.CORE_NODE_TYPE_ATTR + "='LinkLibrary']")
+											&& (m.p === "inside" || m.p === "first" || m.p === "last"))
+										|| (hovered.is("[" + GeoDataTree.CORE_NODE_TYPE_ATTR + "='Link']")
+											&& (m.p == "before" || m.p === "after"))));
+							return valid;
+						}
 					}
 				},
-				plugins: ["ui", "themes", "lazyload", "bettercheckbox", "hoverbuttons", "crr"]
+				plugins: ["ui", "themes", "lazyload", "bettercheckbox", "crrm", "dnd", "hoverbuttons"]
+			})
+			.bind("before.jstree", function (e, data) { 
+				var nms = [], res = true, p, t;
+				if(data.func == "move_node") {
+					// check uniqueness
+					var check_unique = function (nms, p) {
+						var cnms = [];
+						p.find("> a > span > span").each(function () { cnms.push($(this).text().replace(/^\s+/g,"")); });
+						if(!cnms.length || !nms.length) { return true; }
+						cnms = cnms.sort().join(",,").replace(/(,|^)([^,]+)(,,\2)+(,|$)/g,"$1$2$4").replace(/,,+/g,",").replace(/,$/,"").split(",");
+						if((cnms.length + nms.length) != cnms.concat(nms).sort().join(",,").replace(/(,|^)([^,]+)(,,\2)+(,|$)/g,"$1$2$4").replace(/,,+/g,",").replace(/,$/,"").split(",").length) {
+							return false;
+						}
+						return true;
+					};
+					if(data.args[4] === true) {
+						if(data.args[0].o && data.args[0].o.length) {
+							var nms = [], res;
+							data.args[0].o.find("> a > span > span").each(
+									function () {
+										nms.push($(this).text().replace(/^\s+/g,""));
+									}
+								);
+							if (!check_unique(nms, data.args[0].np.find("> ul > li").not(data.args[0].o))) {
+								e.stopPropagation();
+								return false;
+							}
+						}
+					}
+				}
+			})
+			.bind("move_node.jstree", function(e, data) {
+				var linkLibraryGeoData, linkGeoData;
+				linkLibraryGeoData = getGeoDataFromTreeNode(data.args[0].r.closest("li"));
+				linkGeoData = getGeoDataFromTreeNode(data.args[0].o.closest("li"));
+				if (_this.onAppend) {
+					_this.onAppend.call(_this.onAppend, linkLibraryGeoData, 
+							linkGeoData, data.args[0].cp);
+				}				
 			});
 		},
 		
@@ -435,6 +490,8 @@ if (!window.core.ui)
 			var newGeoData = GeoDataStore.getById(geoDataId);
 			var nodeToRefresh = this._findNodeByGeoDataId(geoDataId);
 			this.el.jstree("refresh", nodeToRefresh);
+			console.log(nodeToRefresh);
+			console.log("Refreshed " + nodeToRefresh);
 		}
 
 	};
